requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
addict==2.4.0
    # via open3d
aiohttp==3.8.4
    # via fsspec
aiosignal==1.3.1
    # via aiohttp
ansi2html==1.8.0
    # via dash
asttokens==2.2.1
    # via stack-data
async-timeout==4.0.2
    # via aiohttp
attrdict==2.0.1
    # via -r requirements.in
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
backcall==0.2.0
    # via ipython
certifi==2023.5.7
    # via requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   -r requirements.in
    #   flask
cmake==3.26.4
    # via triton
comm==0.1.3
    # via ipykernel
configargparse==1.5.5
    # via open3d
contourpy==1.1.0
    # via matplotlib
cycler==0.11.0
    # via matplotlib
cython==0.29.36
    # via hdbscan
dash==2.11.1
    # via open3d
dash-core-components==2.0.0
    # via dash
dash-html-components==2.0.0
    # via dash
dash-table==5.0.0
    # via dash
debugpy==1.6.7
    # via ipykernel
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
diskcache==5.4.0
    # via -r requirements.in
executing==1.2.0
    # via stack-data
fastjsonschema==2.17.1
    # via nbformat
filelock==3.12.2
    # via
    #   torch
    #   triton
flask==2.2.5
    # via dash
fonttools==4.41.0
    # via matplotlib
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.6.0
    # via pytorch-lightning
hdbscan==0.8.32
    # via -r requirements.in
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==6.8.0
    # via
    #   flask
    #   jupyter-client
importlib-resources==6.0.0
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   matplotlib
ipdb==0.13.9
    # via -r requirements.in
ipykernel==6.24.0
    # via ipywidgets
ipython==8.12.2
    # via
    #   ipdb
    #   ipykernel
    #   ipywidgets
ipywidgets==8.0.7
    # via open3d
itsdangerous==2.1.2
    # via flask
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via
    #   flask
    #   torch
joblib==1.3.1
    # via
    #   hdbscan
    #   scikit-learn
jsonschema==4.18.4
    # via nbformat
jsonschema-specifications==2023.7.1
    # via jsonschema
jupyter-client==8.3.0
    # via ipykernel
jupyter-core==5.3.1
    # via
    #   ipykernel
    #   jupyter-client
    #   nbformat
jupyterlab-widgets==3.0.8
    # via ipywidgets
keopscore==2.1.2
    # via pykeops
kiwisolver==1.4.4
    # via matplotlib
lightning-utilities==0.9.0
    # via
    #   pytorch-lightning
    #   torchmetrics
lit==16.0.6
    # via triton
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.7.2
    # via open3d
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
nbformat==5.5.0
    # via open3d
nest-asyncio==1.5.6
    # via
    #   dash
    #   ipykernel
networkx==3.1
    # via torch
numpy==1.24.4
    # via
    #   contourpy
    #   hdbscan
    #   matplotlib
    #   open3d
    #   pandas
    #   pykeops
    #   pyquaternion
    #   pytorch-lightning
    #   scikit-learn
    #   scipy
    #   torchmetrics
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   nvidia-cusolver-cu11
    #   torch
nvidia-cuda-cupti-cu11==11.7.101
    # via torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
nvidia-cufft-cu11==10.9.0.58
    # via torch
nvidia-curand-cu11==10.2.10.91
    # via torch
nvidia-cusolver-cu11==11.4.0.1
    # via torch
nvidia-cusparse-cu11==11.7.4.91
    # via torch
nvidia-nccl-cu11==2.14.3
    # via torch
nvidia-nvtx-cu11==11.7.91
    # via torch
open3d==0.16.0
    # via -r requirements.in
packaging==23.1
    # via
    #   ipykernel
    #   lightning-utilities
    #   matplotlib
    #   plotly
    #   pytorch-lightning
    #   torchmetrics
pandas==2.0.3
    # via open3d
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==10.0.0
    # via
    #   matplotlib
    #   open3d
pkgutil-resolve-name==1.3.10
    # via jsonschema
platformdirs==3.9.1
    # via jupyter-core
plotly==5.15.0
    # via dash
prompt-toolkit==3.0.39
    # via ipython
psutil==5.9.5
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pybind11==2.11.1
    # via pykeops
pygments==2.15.1
    # via ipython
pykeops==2.1.2
    # via -r requirements.in
pyparsing==3.0.9
    # via matplotlib
pyquaternion==0.9.9
    # via open3d
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
pytorch-lightning==2.0.5
    # via -r requirements.in
pytz==2023.3
    # via pandas
pyyaml==6.0.1
    # via
    #   open3d
    #   pytorch-lightning
pyzmq==25.1.0
    # via
    #   ipykernel
    #   jupyter-client
referencing==0.30.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   dash
    #   fsspec
retrying==1.3.4
    # via dash
rpds-py==0.9.2
    # via
    #   jsonschema
    #   referencing
scikit-learn==1.3.0
    # via
    #   hdbscan
    #   open3d
scipy==1.10.1
    # via
    #   hdbscan
    #   scikit-learn
six==1.16.0
    # via
    #   asttokens
    #   attrdict
    #   python-dateutil
    #   retrying
stack-data==0.6.2
    # via ipython
sympy==1.12
    # via torch
tenacity==8.2.2
    # via plotly
threadpoolctl==3.2.0
    # via scikit-learn
toml==0.10.2
    # via ipdb
torch==2.0.1
    # via
    #   pytorch-lightning
    #   torchmetrics
    #   triton
torchmetrics==1.0.1
    # via pytorch-lightning
tornado==6.3.2
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.65.0
    # via
    #   open3d
    #   pytorch-lightning
traitlets==5.9.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbformat
triton==2.0.0
    # via torch
typing-extensions==4.7.1
    # via
    #   dash
    #   ipython
    #   lightning-utilities
    #   pytorch-lightning
    #   torch
    #   torchmetrics
tzdata==2023.3
    # via pandas
urllib3==2.0.3
    # via requests
wcwidth==0.2.6
    # via prompt-toolkit
werkzeug==2.2.3
    # via
    #   dash
    #   flask
wheel==0.40.0
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-cupti-cu11
    #   nvidia-cuda-runtime-cu11
    #   nvidia-curand-cu11
    #   nvidia-cusparse-cu11
    #   nvidia-nvtx-cu11
widgetsnbextension==4.0.8
    # via ipywidgets
yarl==1.9.2
    # via aiohttp
zipp==3.16.2
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
